requirements.txt
tensorflow
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
